git-svn-id: svn://localhost/bt1/trunk@28 f571bf5b-3fff-47ac-b1e0-9c329ff6e91c

+++ BrawlLib/Wii/Compressor.cs
@@ -49,7 +49,7 @@ namespace BrawlLib.Wii.Compression
         {
             switch (type)
             {
-                case CompressionType.LZ77: { LZ77.Compact(srcAddr, (uint)srcLen, outStream, null); break; }
+                case CompressionType.LZ77: { LZ77.Compact(srcAddr, srcLen, outStream, null); break; }
             }
         }
     }
